feat: OpenCode v2 plugin compatibility with v1 fallback
- Move original v1 runtime to src/v1/ and keep v1 entrypoints intact
- Add v2 server plugin default export via Plugin.define
- Add hybrid ./tui entry that exposes { id, setup } for v2 while keeping
  v1 tui function accessible through a Proxy
- Pin @opencode-ai/plugin to 0.0.0-beta-17595 and add zod / zod-to-json-schema
- Update package.json exports, build script, and capability manifest
- Add v2 entry tests and fix env isolation in honcho-setup tests
- Update README and CHANGELOG
