+++ src/v1/server.ts
@@ -1,0 +1,10 @@
+import { createHonchoRuntimePlugin } from "./index.js"
+
+export const server = createHonchoRuntimePlugin()
+
+const plugin = {
+  id: "@honcho-ai/opencode-honcho",
+  server,
+}
+
+export default plugin
