+++ tests/server-entry.test.js
@@ -2,7 +2,7 @@ import { expect, test } from "bun:test"
 import { readFile } from "node:fs/promises"
 
 import serverModule from "../dist/server.js"
-import { __testing } from "../dist/index.js"
+import { __testing } from "../dist/v1/index.js"
 
 test("package.json exposes an explicit OpenCode server entry", async () => {
   const pkg = JSON.parse(await readFile(new URL("../package.json", import.meta.url), "utf-8"))
